requirements.txt
annotated-types    0.6.0
anyio              3.7.1
asyncio            3.4.3
certifi            2023.7.22
charset-normalizer 3.3.2
contourpy          1.2.0
cycler             0.12.1
distro             1.8.0
exceptiongroup     1.1.3
fonttools          4.44.0
h11                0.14.0
httpcore           1.0.1
httpx              0.25.1
idna               3.4
kiwisolver         1.4.5
matplotlib         3.8.1
numpy              1.26.1
openai             1.2.2
packaging          23.2
Pillow             10.1.0
pip                23.3
pydantic           2.4.2
pydantic_core      2.10.1
pyparsing          3.1.1
python-dateutil    2.8.2
python-dotenv      1.0.0
requests           2.31.0
setuptools         68.0.0
six                1.16.0
sniffio            1.3.0
tqdm               4.66.1
typing_extensions  4.8.0
urllib3            2.0.7
wheel              0.41.2